Ajout début vérification du membre

+++ controller/TopicController.php
@@ -93,11 +93,13 @@
             
             /* Si le filtrage fonctionne */
             if($id){
-                if($postManager->supprimerPostsDuTopic($id) && $topicManager->delete($id)){ // On supprime les posts du topic puis le topic
-                    $session->addFlash("success", "Suppression réussi !");
-                }
-                else{
-                    $session->addFlash("error", "Echec de la suppression !");
+                if($session->getUser()->getId() == $topicManager->){ // on vérifie que le membre actuelle est bien celui qui supprime le topic
+                    if($postManager->supprimerPostsDuTopic($id) && $topicManager->delete($id)){ // On supprime les posts du topic puis le topic
+                        $session->addFlash("success", "Suppression réussi !");
+                    }
+                    else{
+                        $session->addFlash("error", "Echec de la suppression !");
+                    }
                 }
             }
             else{
